requirements.txt
astroid==2.4.2
certifi==2020.6.20
click==7.1.2
colorama==0.4.3
Flask==1.1.2
gunicorn==20.0.4
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
pylint==2.5.3
toml==0.10.1
Werkzeug==1.0.1
wincertstore==0.2
wrapt==1.12.1
flask_wtf
wtforms
flask_login
flask_sqlalchemy
flask_migrate
Pillow==7.2.0
datetime
flask_table
flask_admin
python-dateutil